requirements.txt
requests
pycrypto

#if you prefer this for DES3 against pycrypto
#pyDes
